requirements.txt
mlx
mlx-lm
pillow
huggingface-hub
mlx-vlm
torch
torchvision
